requirements.txt
-e .
pandas==1.5.3
numpy==1.24.2
openai==1.23.6
langchain==0.1.16
python-dotenv==1.0.1
streamlit==1.33.0
